renamed to .style.ts and moved to src
