+++ src/TypeWriterContainer.style.ts
