+++ src/components/ErrorScreens/OpponentsTeamUnavailable.tsx
